refactor(core/domain): split incident window anchor parsers
Move vendor anchor parsers and ISO-8601 coercion into incident_anchors.py
while keeping IncidentWindow and resolve_incident_window in incident_window.py.
